+++ lib/yt/models/content_owner.rb
@@ -30,6 +30,11 @@ module Yt
       # @return [String] The display name of the content owner.
       attr_reader :display_name
 
+      # @!attribute [r] video_groups
+      #   @return [Yt::Collections::VideoGroups] the video-groups managed by the
+      #     content owner.
+      has_many :video_groups
+
       def initialize(options = {})
         super options
         @owner_name = options[:owner_name]
@@ -55,6 +60,13 @@ module Yt
       # videos *on behalf of* the content owner (public, private, unlisted).
       def videos_params
         {for_content_owner: true, on_behalf_of_content_owner: @owner_name}
+      end
+
+      # @private
+      # Tells `has_many :video_groups` that content_owner.video_groups should
+      # return all the video-groups *on behalf of* the content owner
+      def video_groups_params
+        {on_behalf_of_content_owner: @owner_name}
       end
     end
   end