+++ spec/requests/as_content_owner/content_owner_spec.rb
@@ -51,6 +51,20 @@ describe Yt::ContentOwner, :partner do
     end
   end
 
+  describe '.video_groups' do
+    let(:video_group) { $content_owner.video_groups.first }
+
+    specify 'returns the first video-group created by the account' do
+      expect(video_group).to be_a Yt::VideoGroup
+      expect(video_group.title).to be_a String
+      expect(video_group.item_count).to be_an Integer
+      expect(video_group.published_at).to be_a Time
+    end
+
+    specify 'allows to run reports against each video-group' do
+      expect(video_group.views).to be
+    end
+  end
 
   describe 'claims' do
     let(:asset_id) { ENV['YT_TEST_PARTNER_ASSET_ID'] }